+++ src/main/java/edu/ie3/datamodel/io/factory/result/NodeResultFactory.java
@@ -35,9 +35,7 @@ public class NodeResultFactory extends ResultEntityFactory<NodeResult> {
   @Override
   protected List<Set<String>> getFields(Class<?> entityClass) {
     Set<String> minConstructorParams = newSet(TIME, INPUT_MODEL, VMAG, VANG);
-    Set<String> optionalFields = expandSet(minConstructorParams, ENTITY_UUID);
-
-    return Arrays.asList(minConstructorParams, optionalFields);
+    return List.of(minConstructorParams);
   }
 
   @Override
@@ -47,11 +45,7 @@ public class NodeResultFactory extends ResultEntityFactory<NodeResult> {
     ComparableQuantity<Dimensionless> vMagValue =
         data.getQuantity(VMAG, StandardUnits.VOLTAGE_MAGNITUDE);
     ComparableQuantity<Angle> vAngValue = data.getQuantity(VANG, StandardUnits.VOLTAGE_ANGLE);
-    Optional<UUID> uuidOpt =
-        data.containsKey(ENTITY_UUID) ? Optional.of(data.getUUID(ENTITY_UUID)) : Optional.empty();
 
-    return uuidOpt
-        .map(uuid -> new NodeResult(uuid, zdtTime, inputModelUuid, vMagValue, vAngValue))
-        .orElseGet(() -> new NodeResult(zdtTime, inputModelUuid, vMagValue, vAngValue));
+    return new NodeResult(zdtTime, inputModelUuid, vMagValue, vAngValue);
   }
 }