+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/TimeBasedSimpleValueFactory.java
@@ -18,7 +18,6 @@ import java.util.*;
 
 public class TimeBasedSimpleValueFactory<V extends Value>
     extends TimeBasedValueFactory<SimpleTimeBasedValueData<V>, V> {
-  private static final String UUID = "uuid";
   private static final String TIME = "time";
   /* Energy price */
   private static final String PRICE = "price";
@@ -49,7 +48,6 @@ public class TimeBasedSimpleValueFactory<V extends Value>
 
   @Override
   protected TimeBasedValue<V> buildModel(SimpleTimeBasedValueData<V> data) {
-    UUID uuid = data.getUUID(UUID);
     ZonedDateTime time = timeUtil.toZonedDateTime(data.getField(TIME));
     V value;
 
@@ -83,12 +81,12 @@ public class TimeBasedSimpleValueFactory<V extends Value>
           "The given factory cannot handle target class '" + data.getTargetClass() + "'.");
     }
 
-    return new TimeBasedValue<>(uuid, time, value);
+    return new TimeBasedValue<>(time, value);
   }
 
   @Override
   protected List<Set<String>> getFields(Class<?> entityClass) {
-    Set<String> minConstructorParams = newSet(UUID, TIME);
+    Set<String> minConstructorParams = newSet(TIME);
 
     if (EnergyPriceValue.class.isAssignableFrom(entityClass)) {
       minConstructorParams.add(PRICE);