+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/TimeBasedWeatherValueFactory.java
@@ -16,7 +16,6 @@ import java.util.*;
  */
 public abstract class TimeBasedWeatherValueFactory
     extends TimeBasedValueFactory<TimeBasedWeatherValueData, WeatherValue> {
-  protected static final String UUID = "uuid";
   protected static final String TIME = "time";
   protected static final String COORDINATE_ID = "coordinateId";