+++ src/main/java/edu/ie3/datamodel/models/result/connector/TransformerResult.java
@@ -44,30 +44,6 @@ public abstract class TransformerResult extends ConnectorResult {
     this.tapPos = tapPos;
   }
 
-  /**
-   * @param uuid uuid of this result entity, for automatic uuid generation use primary constructor
-   *     above
-   * @param time date and time when the result is produced
-   * @param inputModel uuid of the input model that produces the result
-   * @param iAMag electric current magnitude @ port A, normally provided in Ampere
-   * @param iAAng electric current angle @ Port A in degree
-   * @param iBMag electric current magnitude @ port B, normally provided in Ampere
-   * @param iBAng electric current angle @ Port B in degree
-   * @param tapPos the current position of the transformers tap changer
-   */
-  protected TransformerResult(
-      UUID uuid,
-      ZonedDateTime time,
-      UUID inputModel,
-      ComparableQuantity<ElectricCurrent> iAMag,
-      ComparableQuantity<Angle> iAAng,
-      ComparableQuantity<ElectricCurrent> iBMag,
-      ComparableQuantity<Angle> iBAng,
-      int tapPos) {
-    super(uuid, time, inputModel, iAMag, iAAng, iBMag, iBAng);
-    this.tapPos = tapPos;
-  }
-
   public int getTapPos() {
     return tapPos;
   }
@@ -93,9 +69,7 @@ public abstract class TransformerResult extends ConnectorResult {
   @Override
   public String toString() {
     return "TransformerResult{"
-        + "uuid="
-        + getUuid()
-        + ", time="
+        + "time="
         + getTime()
         + ", inputModel="
         + getInputModel()