+++ src/main/java/edu/ie3/datamodel/models/timeseries/repetitive/LoadProfileInput.java
@@ -71,9 +71,7 @@ public class LoadProfileInput extends RepetitiveTimeSeries<LoadProfileEntry, PVa
   @Override
   public String toString() {
     return "LoadProfileInput{"
-        + "uuid="
-        + getUuid()
-        + ", type="
+        + "type="
         + type
         + ", dayOfWeekToHourlyValues="
         + dayOfWeekToHourlyValues